refactor amortization file
